Refactor encoder
